[Silver V] Title: 수들의 합, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1789. 수들의 합/수들의 합.cc
@@ -6,7 +6,7 @@ int main()
 {
     cin >> s;
    for(long long i = 1; i<=s; i++){
-       long long num = (long long)i*(i+1)/2;
+       long long num = i*(i+1)/2;
        if(num>s){
            ret = i-1;
            break;
